[Bronze V] Title: 이상한 기호, Time: 56 ms, Memory: 5788 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -4,9 +4,9 @@ namespace Baekjoon
     {
         static void Main()
         {
-            int[] inputs = Console.ReadLine().Split().Select(int.Parse).Take(2).ToArray();
-            int a = inputs[0];
-            int b = inputs[1];
+            long[] inputs = Console.ReadLine().Split().Select(long.Parse).Take(2).ToArray();
+            long a = inputs[0];
+            long b = inputs[1];
 
             Console.WriteLine($"{(a + b) * (a - b)}");
